backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
django-extensions==3.2.3
celery==5.4.0
redis==5.0.1
requests==2.31.0
gunicorn==23.0.0
